+++ crates/core/tedge/src/cli/connect/mod.rs
@@ -1,10 +1,8 @@
-pub use self::c8y_direct_connection::*;
 pub use self::cli::*;
 pub use self::command::*;
 pub use self::error::*;
 
-mod c8y_direct_connection;
+mod c8y;
 mod cli;
 mod command;
 mod error;
-mod jwt_token;
